Fix conda env names in distribution example run template
The Self-Hosted Distribution documentation contain steps
to start llama server via conda environment. If the user
generates conda environment using llaama build command and template,
it generate an environment with name of the distribution and not
defaulyt name of local as per the example run yaml file. It
will thereore fail when user tried to run the server.

This PR fixes the conda env name in the run yaml file.

Signed-off-by: Martin Hickey <martin.hickey@ie.ibm.com>