Add podman instructions to Quick Start
Podman is a popular alternative to Docker, so it would be nice to make it clear that it can also be used to deploy the container for the server.  The instructions are a little different because you have to create the directory (unlike with Docker which makes the directory for you).