Fix logprobs support in remote-vllm provider
The remote-vllm provider was not passing logprobs options from
CompletionRequest or ChatCompletionRequests through to the OpenAI
client parameters. I manually verified this, as well as observed this
provider failing `TestInference::test_completion_logprobs`.

This fixes that by passing the `logprobs.top_k` value through to the
parameters we pass into the OpenAI client.

Additionally, this fixes a bug in `test_text_inference.py` where it
mistakenly assumed chunk.delta were of type `ContentDelta` for
completion requests. The deltas are of type `ContentDelta` for chat
completion requests, but for basic completion requests the deltas are
of type string. This test was likely failing for other providers that
did properly support logprobs because of this latter issue in the
test, which was hit while fixing the above issue with the remote-vllm
provider.

Fixes #1073

Signed-off-by: Ben Browning <bbrownin@redhat.com>