docs: Add note about distro_codegen.py and provider dependencies
This expands upon the existing distro_codegen.py text in the new API
provider documentation to include a note about not including
provider-specific dependencies in the code path that builds the
distribution's template.

Our distro_codegen pre-commit hook will catch this case anyway, but
this attempts to inform provider authors ahead of time about that.

Signed-off-by: Ben Browning <bbrownin@redhat.com>