fix: BuiltinTool JSON serialization in remote vLLM provider
The `tool_name` attribute of `ToolDefinition` instances can either be
a str or a BuiltinTool enum type. This fixes the remote vLLM provider
to use the value of those BuiltinTool enums when serializing to JSON
instead of attempting to serialize the actual enum to JSON.

Reference of how this is handled in some other areas, since I followed
that same pattern for the remote vLLM provider here:
  - [remote nvidia
  provider](https://github.com/meta-llama/llama-stack/blob/v0.1.3/llama_stack/providers/remote/inference/nvidia/openai_utils.py#L137-L140)
  - [meta reference
  provider](https://github.com/meta-llama/llama-stack/blob/v0.1.3/llama_stack/providers/inline/agents/meta_reference/agent_instance.py#L635-L636)

There is opportunity to potentially reconcile the nvidia and
remote-vllm bits where they are both translating Llama Stack Inference
APIs to OpenAI client requests, but that's a can of worms I didn't
want to open for this bug fix.

This explicitly fixes this error when using the remote vLLM provider
and the agent tests:

```
TypeError: Object of type BuiltinTool is not JSON serializable
```

So, this is related to #1144 and addresses the immediate issue raised
there. With this fix,
`tests/client-sdk/agents/test_agents.py::test_builtin_tool_web_search`
now gets past the JSON serialization error when using the remote vLLM
provider and actually attempts to call the web search tool. I don't
have any API keys setup for the actual web search providers yet, so I
cannot verify everything works after that point.

Signed-off-by: Ben Browning <bbrownin@redhat.com>