fix: image_name defaults to none
since `--image-name` has a default of None, often a user can get `llamastack-None` when using environments like venv and container.

currently, the image_name is llamastack- with the value of the flag or the current conda environment. Change this to be either the value of the flag OR the current image_type. So now the default image names are things like:

`llamastack-venv`
`llamastack-conda`
`llamastack-container`

Signed-off-by: Charlie Doern <cdoern@redhat.com>