feat: add inspect API for user-friendly provider config
currently the `inspect` API for providers is really a `list` API. Create a new `inspect/providers/{provider_id}` API
which returns "user friendly" configuration to the end user.

This is done by using an intermediary `UserConfig` Class to look through the current provider configurations and search for fields which contain the following json schema tag: `json_schema_extra={"user_field": True}`.
If a field has this tag, it is added to the returned ProviderConfig.

This enabled users to see their current providers in conjunction with allowed config fields. I have added the ollama URL as an initial field in this paradigm.

This work is part of the RFC at #1359

Signed-off-by: Charlie Doern <cdoern@redhat.com>