fix: Don't confuse CancelledError from implementation shutdown
If implementation raises CancelledError (e.g. when it runs its own async
loop for jobs), the main server shutdown handler gets confused and
doesn't attempt to shut down the main loop tasks. Handle it by
separating exception handling for implementations and for the main loop.

While at it, also fixing the following failure when this happens:

UnboundLocalError: cannot access local variable 'loop' where it is not
associated with a value

Signed-off-by: Ihar Hrachyshka <ihar.hrachyshka@gmail.com>