fix: Fix lifespan app handler not working
It was noticed that shutdown handlers for implementations are not
executed.

Investigation reveals that lifespan logic is not working. When server
starts, we see:

```
ASGI 'lifespan' protocol appears unsupported.
```

This is because uvicorn calls to app with a missing path on startup,
(perhaps to detect that the protocol is working?) and our
TracingMiddleware was not ready for that, resulting in an exception and
lifespan logic disabled.

Fixes #188

Signed-off-by: Ihar Hrachyshka <ihar.hrachyshka@gmail.com>