fix: Enable lifespan=on for uvicorn
Default behavior is "auto", meaning it will try lifespan, but if
something goes wrong, it will silently swallow the issue and won't use
the lifespan handler at all.

This default behavior resulted in silently breaking shutdown handlers
for providers. With this patch, server will fail to start next time
something gets borked in the middleware layer, which is good.

Related: #188

Signed-off-by: Ihar Hrachyshka <ihar.hrachyshka@gmail.com>