feat: Switch synthetic data generation API to jobs pattern
The API should behave in a way similar to existing training and eval
flows where a long running task is sent to background; the client
receives a job ID to follow status and extract artifacts.

Note: there are no providers for this API implemented yet, so no
implementation changes seem to be needed.

Signed-off-by: Ihar Hrachyshka <ihar.hrachyshka@gmail.com>